pyproject.toml
[project]
name = "grocery-agent"
version = "1.0.0"
description = "Grocery agent: recipe ingest, pantry, jumbo.cl cart"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use",
    "fastapi",
    "httpx",
    "jinja2",
    "python-dotenv",
    "trafilatura",
    "uvicorn[standard]",
]

[project.scripts]
start = "grocery_agent.web:run"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["grocery_agent*"]

[tool.uv]
dev-dependencies = []
